# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen Kursus & Perencanaan API",
        "description": "L5 Swagger API documentation for Manajemen Kursus",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Admin - Courses"
                ],
                "summary": "List semua kursus (Admin)",
                "description": "Mendapatkan daftar semua kursus untuk keperluan admin",
                "operationId": "getAdminCourses",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data kursus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Courses"
                ],
                "summary": "Buat kursus baru (Admin)",
                "description": "Menyimpan data kursus baru",
                "operationId": "storeAdminCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Kursus Laravel Dasar"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi kursus"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kursus berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{slug}": {
            "get": {
                "tags": [
                    "Admin - Courses"
                ],
                "summary": "Detail kursus (Admin)",
                "description": "Melihat detail kursus berdasarkan slug",
                "operationId": "showAdminCourse",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kursus berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}": {
            "put": {
                "tags": [
                    "Admin - Courses"
                ],
                "summary": "Update kursus (Admin)",
                "description": "Mengubah data kursus yang ada",
                "operationId": "updateAdminCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Kursus Laravel Update"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi update"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 120000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kursus berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Courses"
                ],
                "summary": "Hapus kursus (Admin)",
                "description": "Menghapus data kursus",
                "operationId": "deleteAdminCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kursus berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}